Android recommendation home

+++ Android/app/src/main/java/com/example/carousel/pojo/Response.kt
@@ -309,7 +309,7 @@ data class ProductDataSearch(
 data class MainProduct(
     @Expose @SerializedName("_id") val _id: String,
     @Expose @SerializedName("title") val title: String,
-    @Expose @SerializedName("rating") val rating: Int,
+    @Expose @SerializedName("rating") val rating: Double,
     @Expose @SerializedName("numberOfRating") val numberOfRating: Int,
 )
 
@@ -328,8 +328,8 @@ data class ResponseVendorMeProduct(
 data class DataProductSearchFilters(
     @Expose @SerializedName("_id") val _id: String,
     @Expose @SerializedName("parameters") val parameters: List<Parameters>,
-    @Expose @SerializedName("maxPrice") val maxPrice: Int,
-    @Expose @SerializedName("minPrice") val minPrice: Int,
+    @Expose @SerializedName("maxPrice") val maxPrice: Double,
+    @Expose @SerializedName("minPrice") val minPrice: Double,
     @Expose @SerializedName("vendors") val vendors: List<VendorID>,
     @Expose @SerializedName("brands") val brands: List<String>,
     @Expose @SerializedName("categories") val categories: List<String>,
